feat: add routes

+++ src/application/use-cases/account-holder/delete-account-holder-use-case.ts
@@ -1,5 +1,4 @@
-import { AccountHolderRepositoryPort } from '@src/ports/account-holder-repository-port';
-import { UseCasePort } from '@src/ports/use-case-port';
+import { AccountHolderRepositoryPort, UseCasePort } from '@src/ports';
 
 export class DeleteAccountHolderUseCase
   implements UseCasePort<DeleteAccountHolderInputDTO, Promise<DeleteAccountHolderOutputDTO>>
@@ -7,6 +6,10 @@ export class DeleteAccountHolderUseCase
   constructor(private readonly accountHolderRepository: AccountHolderRepositoryPort) {}
 
   async execute({ id }: DeleteAccountHolderInputDTO): Promise<void> {
+    const accountHolderExists = await this.accountHolderRepository.getById(id);
+    if (!accountHolderExists) {
+      throw new Error('account holder not found');
+    }
     await this.accountHolderRepository.delete(id);
   }
 }
